[#33] clear unnecessary digraph logic and styles
